requirements.txt
# Core dependencies
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data processing
pandas==2.1.3
pyyaml==6.0.1

# ML & NLP
setfit==0.7.0
transformers==4.36.2
torch>=2.0.0
numpy==1.24.3

# Async & concurrency
aiohttp==3.9.1
asyncio-contextmanager==1.0.0

# Database
sqlalchemy==2.0.23

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
responses==0.14.0

# Code quality
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# Logging & monitoring
python-json-logger==2.0.7

# Optional: Claude API (for TASK 5 fallback)
# anthropic==0.7.6

# Optional: Redis (for TASK 3, 4, 5 caching)
# redis==5.0.1
